Move xml_interface to xmllang
Enables reuse in other driver programs.
It's not CBMC-specific.
